:sparkles: feat: update chat list
