+++ src/components/ui/sheet.tsx
@@ -4,6 +4,7 @@ import * as React from 'react';
 import * as SheetPrimitive from '@radix-ui/react-dialog';
 import { cva, type VariantProps } from 'class-variance-authority';
 import { X } from 'lucide-react';
+import { CgClose } from 'react-icons/cg';
 
 import { cn } from '@/lib/utils';
 
@@ -66,7 +67,7 @@ const SheetContent = React.forwardRef<
     >
       {children}
       <SheetPrimitive.Close className="absolute right-4 top-4 flex items-center justify-center w-7 h-7 p-1.5 rounded-sm opacity-[28%] ring-offset-background transition-opacity hover:opacity-[28%] focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2 disabled:pointer-events-none data-[state=open]:bg-secondary">
-        <X className="h-7 w-7" />
+        <CgClose className="h-7 w-7 p-1 text-[#959595]" />
         <span className="sr-only">Close</span>
       </SheetPrimitive.Close>
     </SheetPrimitive.Content>
