+++ tailwind.config.ts
@@ -126,7 +126,8 @@ const config = {
         xs: { min: '376px' }
       },
       boxShadow: {
-        custom: '0px -2px 8px 0px rgba(31, 30, 30, 0.08)'
+        custom: '0px -2px 8px 0px rgba(31, 30, 30, 0.08)',
+        chatCustom : '0px 1px 2px 0px rgba(0, 0, 0, 0.24), 0px 0px 1px 0px rgba(0, 0, 0, 0.08), 0px 0px 1px 0px rgba(0, 0, 0, 0.08)'
       },
       fontFamily: {
         pretendard: ['var(--font-pretendard)'],
